renaming filter_tasks_view.py --> tasks_filter_view.py
