+++ qtodotxt/ui/views/tasks_filter_view.py
